responsive screen sizes and created test app
